+++ NEWS.md
@@ -1,5 +1,6 @@
 ### Unreleased
 
+* Added texture filtering preference: Auto, Nearest Neighbor, Smooth (#3801)
 * Added 'Collapse All' action and 'Only Expand to Current' mode to Project view (with rhythmcache, #4346)
 * Reduced animated tile marker opacity during terrain editing (by Huy Vũ, #4449)
 * Fixed ability to change properties after deselecting the current object (#4440)